+++ module/simple_exchange.h
@@ -16,43 +16,36 @@
 #include <stdint.h>
 
 /* Exported constants --------------------------------------------------------*/
-#define EXCH_DATA_LENGTH            0x0C    /*!< Msg data array length.       */
-
-#define EXCH_MSG_START              ':'     /*!< Message start of frame flag  */
-#define EXCH_MSG_END                '\n'    /*!< Message end of frame flag    */
-
-#define EXCH_OK                     0x11    /*!< Success operation result     */
-#define EXCH_BUSY                   0x22    /*!< System busy result           */
-#define EXCH_FAIL                   0x33    /*!< Fail operation result        */
+#define EXCH_SOH                    0x01    /*!< Message start of header flag */
 
 /* Exported types ------------------------------------------------------------*/
 /**
-  * @brief  Exchange message structure definition.
-  */
+ * @brief Exchange message structure definition
+ */
 typedef struct
 {
-    uint8_t             start;      /*!< Start of frame symbol.               */
-    uint8_t             cmd;        /*!< Command byte.                        */
-    uint8_t             data[EXCH_DATA_LENGTH];   /*!< Frame data array.      */
-    uint8_t             crc;        /*!< Frame checksum byte.                 */
-    uint8_t             end;        /*!< End of frame symbol.                 */
-
-} EXCH_MsgStruct;
-
-/**
-  * @brief  Exchange union definition for access data as byte array and
-  *         message structure.
-  */
-typedef union
-{
-    EXCH_MsgStruct      fields;
-    uint8_t             all[sizeof(EXCH_MsgStruct)];
+    uint8_t             cmd;        /*!< Command byte                         */
+    uint8_t             *data;      /*!< Frame data array pointer             */
+    uint8_t             length;     /*!< Frame data length                    */
+    uint16_t            crc;        /*!< Frame checksum 2 bytes               */
 
 } EXCH_MsgTypedef;
 
+typedef struct
+{
+    void                (*write_function)(uint8_t);
+    int                 (*read_function)();
+    void                (*parse_function)(EXCH_MsgTypedef*);
+    uint8_t             *buf;
+    uint32_t            length;
+
+} EXCH_InitTypedef;
+
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
-EXCH_MsgTypedef EXCH_PrepareMsg(uint8_t cmd, const uint8_t *data, uint8_t len);
+void EXCH_Init(const EXCH_InitTypedef *exch_init);
+void EXCH_Dispatcher();
+
 uint8_t EXCH_Crc8(const uint8_t *data, uint8_t len);
 uint16_t EXCH_Crc16(const uint8_t *data, uint16_t len);